move things to internal
# Conflicts:
#	internal/handlers.go

# Conflicts:
#	cmd/gcp/main.go
#	ctlog.go

# Conflicts:
#	ctlog.go
