--- timesource.go
@@ -1,32 +1,0 @@
-// Copyright 2016 Google LLC. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package sctfe
-
-import "time"
-
-// timeSource can provide the current time, or be replaced by a mock in tests to return
-// specific values.
-type timeSource interface {
-	// Now returns the current time in real implementations or a suitable value in others
-	Now() time.Time
-}
-
-// SystemTimeSource provides the current system local time
-type SystemTimeSource struct{}
-
-// Now returns the true current local time.
-func (s SystemTimeSource) Now() time.Time {
-	return time.Now()
-}
